Sprite mechanic and more
up to 2 sprites are implemented and only tomato and cheese work.

Butter can be collected and sets the dash number so make the initial dash numbers 0 to work.

Sprite mechanic works on a namespace basis where it checks the appropriate sprite to modify and the appropriate sprite to give it